Move xml views to event collections in order to have occurrences nested under event
