+++ src/onegov/event/collections/events.py
@@ -5,8 +5,10 @@ from collections import namedtuple
 from datetime import date, timezone
 from datetime import datetime
 from datetime import timedelta
+
 from icalendar import Calendar as vCalendar
 from icalendar.prop import vCategory
+from lxml import objectify, etree
 from markupsafe import escape
 
 from onegov.core.collection import Pagination
@@ -423,3 +425,223 @@ class EventCollection(Pagination):
 
         return self.from_import(items, publish_immediately=True,
                                 future_events_only=future_events_only)
+
+    def as_xml(self, future_events_only=True):
+        """
+        Returns all published occurrences as xml.
+
+        The xml format was Winterthur's wish (no specs behind). Their mobile
+        app will consume the events from xml
+
+        Format:
+        <events>
+            <event>
+                <id></id>
+                <title></title>
+                <tags></tags>
+                    <tag></tag>
+                <description></description>
+                <start></start>
+                <end></end>
+                <location></location>
+                <price></price>
+                ..
+            </event>
+            <event>
+                ..
+            </event>
+            ..
+        </events>
+
+        :param future_events_only: if set, only future events will be
+        returned, all events otherwise
+        :rtype: str
+        :return: xml string
+
+        """
+        xml = '<events></events>'
+        root = objectify.fromstring(xml)
+
+        query = self.session.query(Event)
+        for e in query:
+            # e = (self.session.query(Event)
+            #     .filter(Event.id == e.event_id).first())
+
+            if e.state != 'published':
+                continue
+            if future_events_only and datetime.fromisoformat(str(
+                    e.end)).date() < datetime.today().date():
+                continue
+
+            event = objectify.Element('event')
+            event.id = e.id
+            event.title = e.title
+            txs = tags(e.tags)
+            event.append(txs)
+            event.description = e.description
+            event.start = e.start
+            event.end = e.end
+            event.location = e.location
+            event.price = e.price
+            event.organizer = e.organizer
+            event.event_url = e.external_event_url
+            event.organizer_email = e.organizer_email
+            event.organizer_phone = e.organizer_phone
+            event.modified = e.last_change
+            root.append(event)
+
+        # remove lxml annotations
+        objectify.deannotate(root, pytype=True, xsi=True, xsi_nil=True)
+        etree.cleanup_namespaces(root)
+
+        return etree.tostring(root, encoding='utf-8', xml_declaration=True,
+                              pretty_print=True)
+
+    def as_anthrazit_xml(self, request, future_events_only=True):
+        """
+        Returns all published occurrences as xml for Winterthur.
+        Anthrazit format according
+        https://doc.anthrazit.org/ext/XML_Schnittstelle
+
+        <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+        <import partner="???" partnerid"???" passwort"???" importid="??">
+            <item status="1" suchbar="1" mutationsdatum="2023-08-18 08:23:30">
+                <id>01</id>
+                <titel>Titel der Seite</titel>
+                <textmobile>2-3 Sätze des Text Feldes</textmobile>
+                <termin allday="1">
+                    <von>2011-08-06 00:00:00</von>
+                    <bis>2011-08-06 23:59:00</bis>
+                </termin>
+                <text>Beschreibung</text>
+                <urlweb>url</urlweb>
+                <hauptrubrik name="Naturmusuem">
+                    <rubrik>tag_1</rubrik>
+                    <rubrik>tag_2</rubrik>
+                </hauptrubrik>
+                <keyword>Naturmusuem tag_1 tag_2</keyword>
+                <veranstaltungsort>
+                    <title></title>
+                    <adresse></adresse>
+                    <plz></plz>
+                    <ort></ort>
+                </veranstaltungsort>
+                ...
+            </item>
+            <item>
+                ...
+            </item>
+        </import>
+
+        :param future_events_only: if set, only future events will be
+        returned, all events otherwise
+        :rtype: str
+        :return: xml string
+
+        """
+        cdata = '<![CDATA[{}]]>'
+
+        xml = ('<import partner="" partnerid="" passwort="" importid="">'
+               '</import>')
+        root = etree.XML(xml)
+
+        query = self.session.query(Event)
+        for e in query:
+            if e.state != 'published':
+                continue
+            if future_events_only and datetime.fromisoformat(str(
+                    e.end)).date() < datetime.today().date():
+                continue
+
+            # TODO translate tags
+            last_change = e.last_change.strftime('%Y-%m-%d %H:%M:%S')
+            event = objectify.Element(
+                'item',
+                {
+                    'status': '1',
+                    'suchbar': '1',
+                    'mutationsdatum': last_change
+                }
+            )
+            event.id = e.id
+            event.titel = e.title
+            if len(e.description) > 100:
+                event.textmobile = e.description[:100] + '..'
+            else:
+                event.textmobile = e.description
+            for occ in e.occurrences:
+                termin = objectify.Element('termin')
+                termin.von = occ.localized_start
+                termin.bis = occ.localized_end
+                event.append(termin)
+
+            if e.price:
+                event.sf01 = cdata.format(e.price)
+
+            event.append(text_tag(e.description))
+            if e.external_event_url:
+                event.url_web = e.external_event_url
+            if e.image:
+                event.url_bild = request.link(e.image)
+
+            hr_text = ''
+            tags = []
+            if e.tags:
+                tags = e.tags
+            if e.filter_keywords:
+                for k, v in e.filter_keywords.items():
+                    if k in ['kalender']:
+                        hr_text = v
+                    else:
+                        if isinstance(v, list):
+                            tags.extend(v)
+                        else:
+                            tags.append(v)
+            hr = objectify.Element('hauptrubrik',
+                                   attrib={'name': hr_text} if
+                                   hr_text else None)
+            hr.rubrik = tags or None
+            event.append(hr)
+
+            ort = objectify.Element('veranstaltungsort')
+            ort.titel = e.location
+            if e.coordinates:
+                ort.longitude = e.coordinates.lon
+                ort.latitude = e.coordinates.lat
+            event.append(ort)
+            root.append(event)
+
+        # remove lxml annotations
+        objectify.deannotate(root, pytype=True, xsi=True, xsi_nil=True,
+                             cleanup_namespaces=True)
+
+        return etree.tostring(root, encoding='utf-8', xml_declaration=True,
+                              pretty_print=True)
+
+
+class tags(etree.ElementBase):
+    """
+    Custom class as 'tag' is a member of class Element and cannot be
+    used as tag name.
+    """
+
+    def __init__(self, tags=()):
+        super().__init__()
+        self.tag = 'tags'
+
+        for t in tags:
+            tag = etree.Element('tag')
+            tag.text = t
+            self.append(tag)
+
+
+class text_tag(etree.ElementBase):
+    """
+    Custom class as 'text' is a member of class Element and cannot be
+    used as tag name.
+    """
+
+    def __init__(self, text):
+        super().__init__()
+        self.tag = 'text'
+        self.text = text
