+++ src/onegov/org/views/event.py
@@ -455,6 +455,25 @@ def ical_export_event(self, request):
 
 
 @OrgApp.view(
+    model=OccurrenceCollection,
+    name='xml',
+    permission=Public
+)
+def xml_export_all_events(self, request):
+    """
+    Returns events as xml.
+    Url for xml view: ../events/xml
+
+    """
+    collection = EventCollection(request.session)
+    return Response(
+        collection.as_xml(),
+        content_type='text/xml',
+        content_disposition='inline; filename=events.xml'
+    )
+
+
+@OrgApp.view(
     model=Event,
     name='latest',
     permission=Public
