+++ src/onegov/winterthur/views/occurrence.py
@@ -3,12 +3,12 @@
 from morepath.request import Response
 
 from onegov.core.security import Public
-from onegov.event import OccurrenceCollection
+from onegov.event import OccurrenceCollection, EventCollection
 from onegov.winterthur import WinterthurApp
 
 
 @WinterthurApp.view(model=OccurrenceCollection, name='xml', permission=Public)
-def xml_export_all_occurrences(self, request):
+def xml_export_all_events(self, request):
     """
     Returns events as xml in Anthrazit format.
     This view was requested by Winterthur for their mobile app that displays
@@ -17,9 +17,9 @@ def xml_export_all_occurrences(self, request):
     Url for xml view: ../events/xml
 
     """
-    collection = OccurrenceCollection(request.session)
+    collection = EventCollection(request.session)
     return Response(
-        collection.as_anthrazit_xml(),
+        collection.as_anthrazit_xml(request),
         content_type='text/xml',
-        content_disposition='inline; filename=occurrences.xml'
+        content_disposition='inline; filename=events.xml'
     )
